requirements.txt
black==24.4.2
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
flask==3.0.3
flask-session==0.8.0
greenlet==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
tomli==2.0.1
typing-extensions==4.12.2
werkzeug==3.0.3
zipp==3.19.2
